requirements.txt
beautifulsoup4==4.12.3
certifi==2024.6.2
Flask==3.0.3
Flask_Cors==4.0.1
openai==1.36.1
pymongo==4.8.0
pytest==8.3.1
python-socketio==5.11.3
Requests==2.32.3
